channelsales/requirements.txt
django==1.11.11
djangorestframework==3.8.2
psycopg2